fix math

+++ libraries/OBD/OBD.h
@@ -184,7 +184,7 @@ protected:
 private:
 	virtual uint8_t getPercentageValue(char* data)
 	{
-		return (uint8_t)((float)hex2uint8(data) * 100.0f / 255.0f);
+		return (uint16_t)hex2uint8(data) * 100 / 255;
 	}
 	virtual uint16_t getLargeValue(char* data)
 	{
